working copy

+++ delete-files.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 
 usage() {
-    echo "Usage: $0 --path <path_to_start_from> --target <target_name> [--dry]"
+    echo "Usage: $0 --path <path_to_start_from> --target <target_name> [--dry] [--recycle-bin-path <path_to_recycle_bin>]"
     exit 1
 }
 
 DRY_RUN=0
 TARGET_PATH=""
 TARGET_NAME=""
+RECYCLE_BIN_PATH=""
 
 while [[ $# -gt 0 ]]; do
     key="$1"
@@ -26,6 +27,11 @@ while [[ $# -gt 0 ]]; do
         DRY_RUN=1
         shift # past argument
         ;;
+        --recycle-bin-path)
+        RECYCLE_BIN_PATH="$2"
+        shift # past argument
+        shift # past value
+        ;;
         *)
         usage
         ;;
@@ -43,13 +49,13 @@ else
     exit 1
 fi
 
-
 # Function to search and delete (or dry run)
 
 function delete_target {
     local path="$1"
     local targets="$2"
     local is_dry_run="$3"
+    local recycle_bin="$4"
 
     IFS=',' read -ra TARGET_ARRAY <<< "$targets"
 
@@ -59,12 +65,30 @@ function delete_target {
                 if [[ $is_dry_run -eq 1 ]]; then
                     echo "Will delete: $line"
                 else
-                    if [[ -d "$line" ]]; then
-                        echo "Attempting to delete directory: $line"
-                        rmdir "$line" 2>/dev/null || echo "Skipping non-empty directory: $line"
-                    elif [[ -f "$line" ]]; then
-                        echo "Deleting file: $line"
-                        rm -f "$line"
+                    if [[ -n "$recycle_bin" ]]; then
+                        local recycle_path="${recycle_bin}${line#$path}"
+
+                        # Check if the line is a directory; if so, adjust the path
+                        if [[ -d "$line" ]]; then
+                            recycle_path="${recycle_bin}${line#$path}"
+                        fi
+                        
+                        # Create the directory only if it does not exist
+                        local recycle_dir="$(dirname "$recycle_path")"
+                        if [[ ! -d "$recycle_dir" ]]; then
+                            mkdir -p "$recycle_dir"
+                        fi
+
+                        echo "Moving to recycle bin: $recycle_path"
+                        mv "$line" "$recycle_path"
+                    else
+                        if [[ -d "$line" ]]; then
+                            echo "Attempting to delete directory: $line"
+                            rmdir "$line" 2>/dev/null || echo "Skipping non-empty directory: $line"
+                        elif [[ -f "$line" ]]; then
+                            echo "Deleting file: $line"
+                            rm -f "$line"
+                        fi
                     fi
                 fi
             fi
@@ -72,8 +96,5 @@ function delete_target {
     done
 }
 
-
-# }
-
 # Call the function
-delete_target "$TARGET_PATH" "$TARGET_NAME" "$DRY_RUN"
+delete_target "$TARGET_PATH" "$TARGET_NAME" "$DRY_RUN" "$RECYCLE_BIN_PATH"
